@@ -1,20 +1,24 @@
 import 'dart:convert';
 import 'dart:math';
 
-import 'package:dio/dio.dart';
-import 'package:firebase_analytics/firebase_analytics.dart';
-import 'package:firebase_remote_config/firebase_remote_config.dart';
-import 'package:flare_flutter/flare_actor.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
+
+import 'package:dio/dio.dart';
+import 'package:firebase_analytics/firebase_analytics.dart';
+import 'package:firebase_remote_config/firebase_remote_config.dart';
+import 'package:flare_flutter/flare_actor.dart';
 import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:get/get.dart';
+import 'package:url_launcher/url_launcher.dart';
+import 'package:video_player/video_player.dart';
+
 import 'package:mi_utem/models/usuario.dart';
 import 'package:mi_utem/screens/main_screen.dart';
-import 'package:mi_utem/services/autenticacion_service.dart';
+import 'package:mi_utem/services/auth_service.dart';
 import 'package:mi_utem/services/config_service.dart';
 import 'package:mi_utem/themes/theme.dart';
 import 'package:mi_utem/widgets/acerca_dialog.dart';
@@ -22,9 +26,8 @@ import 'package:mi_utem/widgets/acerca_screen.dart';
 import 'package:mi_utem/widgets/error_dialog.dart';
 import 'package:mi_utem/widgets/loading_dialog.dart';
 import 'package:mi_utem/widgets/login_text_form_field.dart';
-import 'package:url_launcher/url_launcher.dart';
+
 //import 'package:new_version/new_version.dart';
-import 'package:video_player/video_player.dart';
 
 class LoginScreen extends StatefulWidget {
   LoginScreen({Key? key}) : super(key: key);
@@ -66,7 +69,7 @@ class _LoginScreenState extends State<LoginScreen> {
         setState(() {});
       });
 
-    SchedulerBinding.instance!.addPostFrameCallback((_) {
+    SchedulerBinding.instance.addPostFrameCallback((_) {
       _checkAndPerformUpdate();
     });
   }
@@ -322,9 +325,9 @@ class _LoginScreenState extends State<LoginScreen> {
         );
 
         try {
-          bool esPrimeraVez = await AutenticacionService.esPrimeraVez();
+          bool esPrimeraVez = await AuthService.esPrimeraVez();
           Usuario usuario =
-              await AutenticacionService.login(_correo, _contrasenia, true);
+              await AuthService.login(_correo, _contrasenia, true);
 
           FirebaseAnalytics.instance.logLogin();
           FirebaseAnalytics.instance.setUserId(id: usuario.correoUtem);
@@ -379,7 +382,11 @@ class _LoginScreenState extends State<LoginScreen> {
                           ),
                           recognizer: TapGestureRecognizer()
                             ..onTap = () {
-                              launch("https://mi.utem.cl/?ref=appMiUtemInndev");
+                              launchUrl(
+                                Uri.parse(
+                                  "https://mi.utem.cl/?ref=appMiUtemInndev",
+                                ),
+                              );
                             },
                         ),
                         TextSpan(text: ", contáctate con "),
@@ -390,7 +397,11 @@ class _LoginScreenState extends State<LoginScreen> {
                           ),
                           recognizer: TapGestureRecognizer()
                             ..onTap = () {
-                              launch("mailto:soporte.sisei@utem.cl");
+                              launchUrl(
+                                Uri.parse(
+                                  "mailto:soporte.sisei@utem.cl",
+                                ),
+                              );
                             },
                         ),
                         TextSpan(
