@@ -1,11 +1,13 @@
 import 'dart:typed_data';
 
+import 'package:flutter/material.dart';
+
 import 'package:barcode_image/barcode_image.dart';
 import 'package:barcode_widget/barcode_widget.dart';
-import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:image/image.dart' as dartImage;
 import 'package:intl/intl.dart';
+
 import 'package:mi_utem/models/permiso_covid.dart';
 import 'package:mi_utem/models/usuario.dart';
 import 'package:mi_utem/services/permisos_covid_service.dart';
@@ -74,6 +76,7 @@ class LoadedScreen extends StatelessWidget {
 
   _openQr(String heroTag) {
     final image = dartImage.Image(500, 500);
+
     dartImage.fill(image, dartImage.getColor(255, 255, 255));
     drawBarcode(
       image,
