@@ -1,13 +1,14 @@
 import 'package:dio/dio.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:get_storage/get_storage.dart';
+
 import 'package:mi_utem/models/carrera.dart';
 import 'package:mi_utem/models/usuario.dart';
 import 'package:mi_utem/services/carreras_service.dart';
 import 'package:mi_utem/services/perfil_service.dart';
 import 'package:mi_utem/utils/dio_miutem_client.dart';
 
-class AutenticacionService {
+class AuthService {
   static final Dio _dio = DioMiUtemClient.initDio;
   static final GetStorage box = GetStorage();
 
@@ -125,5 +126,46 @@ class AutenticacionService {
       print(e.toString());
       throw e;
     }
+  }
+
+  static Future<String> refreshToken() async {
+    String uri = "/v1/auth";
+
+    try {
+      final FlutterSecureStorage secureStorage = new FlutterSecureStorage();
+
+      String? correo = box.read("correoUtem");
+      String? contrasenia = await secureStorage.read(key: "contrasenia");
+
+      if (correo != null && contrasenia != null) {
+        dynamic data = {'correo': correo, 'contrasenia': contrasenia};
+
+        Response response = await DioMiUtemClient.initDio.post(uri, data: data);
+
+        Usuario usuario = Usuario.fromJson(response.data);
+        _storeToken(usuario.token!);
+
+        return usuario.token!;
+      }
+      throw Exception("No se pudo refrescar el token");
+    } catch (e) {
+      print(e.toString());
+      rethrow;
+    }
+  }
+
+  static void _storeToken(String token) async {
+    box.write('token', token);
+  }
+
+  static String getToken() {
+    final token = box.read('token');
+    if (token == null) throw Exception("No se ha encontrado el token");
+
+    return token;
+  }
+
+  static void invalidateToken() async {
+    box.remove('token');
   }
 }
