@@ -1,11 +1,8 @@
-import 'dart:convert';
 import 'package:dio/dio.dart';
-import 'package:mi_utem/models/rut.dart';
+import 'package:http/http.dart' as http;
+
 import 'package:mi_utem/models/usuario.dart';
 import 'package:mi_utem/utils/dio_docente_client.dart';
-import 'package:mi_utem/utils/dio_miutem_client.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-import 'package:http/http.dart' as http;
 
 class DocentesService {
   static final Dio _dio = DioDocenteClient.initDio;
@@ -18,8 +15,7 @@ class DocentesService {
 
     for (var formato in formatos) {
       String actualImageUrl = "$baseUrl${usuario.rut!.numero}$formato";
-      final imageResponse =
-        await http.head(Uri.parse(actualImageUrl));
+      final imageResponse = await http.head(Uri.parse(actualImageUrl));
 
       if (imageResponse.statusCode == 200) {
         imageUrl = actualImageUrl;
@@ -32,7 +28,7 @@ class DocentesService {
 
   static Future<List<Usuario>> buscarDocentes(String nombre) async {
     String uri = "/docentes/buscar";
-    
+
     try {
       dynamic data = {"nombre": nombre};
 
@@ -41,9 +37,8 @@ class DocentesService {
       List<Usuario> usuarios = Usuario.fromJsonList(response.data["docentes"]);
       List<Usuario> usuariosConFoto = [];
       for (var usuario in usuarios) {
-        
         Usuario usuarioConFoto = usuario;
-        
+
         usuarioConFoto.fotoUrl = await generarImagenPerfil(usuario);
         usuariosConFoto.add(usuarioConFoto);
       }
@@ -56,7 +51,7 @@ class DocentesService {
 
   static Future<Usuario> traerUnDocente(String? nombre) async {
     String uri = "/docentes/buscar";
-    
+
     try {
       dynamic data = {"nombre": nombre, "limit": 1};
 
@@ -66,7 +61,7 @@ class DocentesService {
 
       Usuario usuarioConFoto = usuario;
       usuarioConFoto.fotoUrl = await generarImagenPerfil(usuario);
-      
+
       return usuarioConFoto;
     } on DioError catch (e) {
       print(e.message);
@@ -74,11 +69,16 @@ class DocentesService {
     }
   }
 
-  static Future<Usuario> asignarUnDocente(String? nombreDocente, String? codigoAsignatura, String? nombreAsignatura) async {
+  static Future<Usuario> asignarUnDocente(String? nombreDocente,
+      String? codigoAsignatura, String? nombreAsignatura) async {
     String uri = "/docentes/asignar";
-    
+
     try {
-      dynamic data = {"nombreDocente": nombreDocente, "codigoAsignatura": codigoAsignatura, "nombreAsignatura": nombreAsignatura};
+      dynamic data = {
+        "nombreDocente": nombreDocente,
+        "codigoAsignatura": codigoAsignatura,
+        "nombreAsignatura": nombreAsignatura
+      };
 
       Response response = await _dio.post(uri, data: data);
 
@@ -86,12 +86,11 @@ class DocentesService {
 
       Usuario usuarioConFoto = usuario;
       usuarioConFoto.fotoUrl = await generarImagenPerfil(usuario);
-      
+
       return usuarioConFoto;
     } on DioError catch (e) {
       print(e.message);
       throw e;
     }
   }
-
 }
