@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 class MainTheme {
   static Color get utemAzul => Color(0xff06607a);
@@ -35,14 +36,14 @@ class MainTheme {
                 const Set<MaterialState> interactiveStates = <MaterialState>{
                   MaterialState.disabled,
                 };
-                if (states.any(interactiveStates.contains)) {
-                  return disabledColor;
-                }
-                return primaryColor;
+                return states.any(interactiveStates.contains)
+                    ? disabledColor
+                    : primaryColor;
               },
             ),
             padding: MaterialStateProperty.all(
-                EdgeInsets.symmetric(vertical: 5, horizontal: 20)),
+              EdgeInsets.symmetric(vertical: 5, horizontal: 20),
+            ),
             foregroundColor: MaterialStateProperty.all(Colors.white),
             shape: MaterialStateProperty.all(
               StadiumBorder(),
@@ -57,25 +58,21 @@ class MainTheme {
                 const Set<MaterialState> interactiveStates = <MaterialState>{
                   MaterialState.disabled,
                 };
-                if (states.any(interactiveStates.contains)) {
-                  return BorderSide(
-                    color: disabledColor!,
-                  );
-                }
-                return BorderSide(
-                  color: primaryColor,
-                );
+                return states.any(interactiveStates.contains)
+                    ? BorderSide(color: disabledColor!)
+                    : BorderSide(color: primaryColor);
               },
             ),
-            foregroundColor: MaterialStateProperty.resolveWith((states) {
-              const Set<MaterialState> interactiveStates = <MaterialState>{
-                MaterialState.disabled,
-              };
-              if (states.any(interactiveStates.contains)) {
-                return disabledColor;
-              }
-              return primaryColor;
-            }),
+            foregroundColor: MaterialStateProperty.resolveWith(
+              (states) {
+                const Set<MaterialState> interactiveStates = <MaterialState>{
+                  MaterialState.disabled,
+                };
+                return states.any(interactiveStates.contains)
+                    ? disabledColor
+                    : primaryColor;
+              },
+            ),
             padding: MaterialStateProperty.all(
                 EdgeInsets.symmetric(vertical: 5, horizontal: 20)),
             shape: MaterialStateProperty.all(
@@ -109,15 +106,18 @@ class MainTheme {
           color: primaryColor,
           elevation: 0,
           iconTheme: IconThemeData(color: Colors.white),
-          brightness: Brightness.dark,
+          systemOverlayStyle: SystemUiOverlayStyle.light,
         ),
         inputDecorationTheme: InputDecorationTheme(
-            contentPadding: EdgeInsets.symmetric(vertical: 10, horizontal: 10),
-            isDense: true,
-            hintStyle: TextStyle(color: Colors.grey[400]),
-            border: OutlineInputBorder(
-              borderRadius: BorderRadius.all(Radius.circular(15)),
-            )),
+          contentPadding: EdgeInsets.symmetric(vertical: 10, horizontal: 10),
+          isDense: true,
+          hintStyle: TextStyle(color: Colors.grey[400]),
+          border: OutlineInputBorder(
+            borderRadius: BorderRadius.all(
+              Radius.circular(15),
+            ),
+          ),
+        ),
         primaryColor: primaryColor,
         accentColor: primaryLightColor,
       );
