@@ -1,5 +1,4 @@
 import 'package:dio/dio.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 class DioWordpressClient {
   static const String url = 'https://www.utem.cl/wp-json/wp/v2';
@@ -7,8 +6,8 @@ class DioWordpressClient {
   static Dio _dio = Dio(BaseOptions(
     baseUrl: url,
   ));
-  
+
   static Dio get initDio => _dio
-    //..interceptors.add(DioCacheManager(CacheConfig(baseUrl: url)).interceptor)
-    ;
+      //..interceptors.add(DioCacheManager(CacheConfig(baseUrl: url)).interceptor)
+      ;
 }
