@@ -1,7 +1,9 @@
 import 'dart:math';
 
-import 'package:charts_flutter/flutter.dart' as charts;
 import 'package:flutter/material.dart';
+
+import 'package:charts_flutter/flutter.dart' as charts;
+
 import 'package:mi_utem/models/asignatura.dart';
 import 'package:mi_utem/themes/theme.dart';
 
