@@ -1,5 +1,5 @@
-import 'package:flare_flutter/flare_actor.dart';
 import 'package:flutter/material.dart';
+
 import 'package:get/get.dart';
 
 final _formKey = GlobalKey<FormState>();
@@ -14,48 +14,36 @@ class ErrorDialog extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Dialog(     
-      elevation: 0,
-      backgroundColor: Colors.transparent,
-      child: Container(
-        
-        child: Card(
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(15.0),
-          ),
-          child: Padding( 
-            padding: EdgeInsets.all(30),
-            child: Form(
-              key: _formKey,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                crossAxisAlignment: CrossAxisAlignment.center,
-                mainAxisSize: MainAxisSize.min,
-                children: <Widget>[
-                  Container(
-                    child: this.mensaje
-                  ),
-                  if (this.contenido != null)
-                    Padding(
-                        padding: EdgeInsets.all(20),
-                        child: this.contenido
-                      ),
-                  Padding(
-                    padding: EdgeInsets.only(top: 20),
-                    child: TextButton(
-                      onPressed: () async {
-                        Get.back();
-                      },
-                      child: Text("Entendido 😥"),
-                    )
-                  )
-                  
-                ]
-              )
-            )
-          )
-        )
-      )
-    );
+    return Dialog(
+        elevation: 0,
+        backgroundColor: Colors.transparent,
+        child: Container(
+            child: Card(
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(15.0),
+                ),
+                child: Padding(
+                    padding: EdgeInsets.all(30),
+                    child: Form(
+                        key: _formKey,
+                        child: Column(
+                            mainAxisAlignment: MainAxisAlignment.center,
+                            crossAxisAlignment: CrossAxisAlignment.center,
+                            mainAxisSize: MainAxisSize.min,
+                            children: <Widget>[
+                              Container(child: this.mensaje),
+                              if (this.contenido != null)
+                                Padding(
+                                    padding: EdgeInsets.all(20),
+                                    child: this.contenido),
+                              Padding(
+                                  padding: EdgeInsets.only(top: 20),
+                                  child: TextButton(
+                                    onPressed: () async {
+                                      Get.back();
+                                    },
+                                    child: Text("Entendido 😥"),
+                                  ))
+                            ]))))));
   }
 }
