@@ -1,10 +1,9 @@
 import 'dart:async';
 import 'dart:typed_data';
-import 'dart:ui';
+
+import 'package:flutter/material.dart';
 
 import 'package:extended_image/extended_image.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:image_editor/image_editor.dart';
 
@@ -141,8 +140,6 @@ class _ImagenEditorModalState extends State<ImagenEditorModal> {
     EditActionDetails action = editorKey.currentState!.editAction!;
 
     final Rect? rect = state.getCropRect();
-
-    Uint8List data = state.rawImageData;
 
     final rotateAngle = action.rotateAngle.toInt();
     final flipHorizontal = action.flipY;
